Move into using the more structured Provider/Hooks/Assets class structure.
